♻️ Refactor: strongly-type Gcal and SyncErrors
