+++ packages/backend/src/common/constants/error.constants.ts
@@ -15,44 +15,6 @@ export const SocketError: ErrorConstant = {
   },
 };
 
-export const SyncError: ErrorConstant = {
-  AccessRevoked: {
-    description: "User revoked access to their 3rd-party calendar (GCal)",
-    status: Status.GONE,
-    isOperational: true,
-  },
-  CalendarWatchExists: {
-    description: "Watch already exists",
-    status: Status.BAD_REQUEST,
-    isOperational: true,
-  },
-  NoGCalendarId: {
-    description: "No gCalendarId",
-    status: Status.NO_CONTENT,
-    isOperational: true,
-  },
-  NoResourceId: {
-    description: "No resourceId provided",
-    status: Status.NO_CONTENT,
-    isOperational: true,
-  },
-  NoSyncToken: {
-    description: "No syncToken",
-    status: Status.INTERNAL_SERVER,
-    isOperational: true,
-  },
-  NoEventChanges: {
-    description: "Nothing changed",
-    status: Status.BAD_REQUEST,
-    isOperational: true,
-  },
-  NoSyncRecordForUser: {
-    description: "No sync record for user",
-    status: Status.BAD_REQUEST,
-    isOperational: true,
-  },
-};
-
 export const UserError: ErrorConstant = {
   InvalidValue: {
     description: "User has an invalid value",
