+++ packages/backend/src/common/errors/sync/sync.errors.ts
@@ -1,4 +1,50 @@
 import { Status } from "@core/errors/status.codes";
 import { ErrorMetadata } from "@backend/common/types/error.types";
 
-interface SyncErrors {}
+interface SyncErrors {
+  AccessRevoked: ErrorMetadata;
+  CalendarWatchExists: ErrorMetadata;
+  NoGCalendarId: ErrorMetadata;
+  NoResourceId: ErrorMetadata;
+  NoSyncToken: ErrorMetadata;
+  NoEventChanges: ErrorMetadata;
+  NoSyncRecordForUser: ErrorMetadata;
+}
+
+export const SyncError: SyncErrors = {
+  AccessRevoked: {
+    description: "User revoked access to their 3rd-party calendar (GCal)",
+    status: Status.GONE,
+    isOperational: true,
+  },
+  CalendarWatchExists: {
+    description: "Watch already exists",
+    status: Status.BAD_REQUEST,
+    isOperational: true,
+  },
+  NoGCalendarId: {
+    description: "No gCalendarId",
+    status: Status.NO_CONTENT,
+    isOperational: true,
+  },
+  NoResourceId: {
+    description: "No resourceId provided",
+    status: Status.NO_CONTENT,
+    isOperational: true,
+  },
+  NoSyncToken: {
+    description: "No syncToken",
+    status: Status.INTERNAL_SERVER,
+    isOperational: true,
+  },
+  NoEventChanges: {
+    description: "Nothing changed",
+    status: Status.BAD_REQUEST,
+    isOperational: true,
+  },
+  NoSyncRecordForUser: {
+    description: "No sync record for user",
+    status: Status.BAD_REQUEST,
+    isOperational: true,
+  },
+};
