+++ packages/backend/src/sync/services/notify/handler/gcal.notification.handler.test.ts
@@ -4,8 +4,8 @@ import {
   setupTestDb,
 } from "@backend/__tests__/helpers/mock.db.setup";
 import { mockRegularGcalEvent } from "@backend/__tests__/mocks.gcal/factories/gcal.event.factory";
-import { SyncError } from "@backend/common/constants/error.constants";
 import { error } from "@backend/common/errors/handlers/error.handler";
+import { SyncError } from "@backend/common/errors/sync/sync.errors";
 import gcalService from "@backend/common/services/gcal/gcal.service";
 import { getSync } from "@backend/sync/util/sync.queries";
 import { GCalNotificationHandler } from "./gcal.notification.handler";
